requirement.txt
asgiref==3.8.1
black==22.12.0
click==8.1.3
Django==5.1.2
django-cors-headers==4.3.1
django-filter==22.1
django-rest-passwordreset==1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.8
flake8==6.0.0
inflection==0.5.1
isort==5.11.4
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==24.0
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
psycopg2-binary==2.9.10
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-decouple==3.8
pytz==2022.7
PyYAML==6.0.1
sqlparse==0.5.1
tomli==2.0.1
uritemplate==4.1.1
